# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Автоматизированная система управления требованиями (АСУТр)
  description: |
    API для системы управления требованиями согласно ТЗ ГОСТ 34.602-2020.
    Система предоставляет функционал для управления требованиями, их версионированием,
    согласованием, трассировкой и интеграцией со смежными системами.
  version: 1.0.0
  contact:
    name: АО "МИФИСТ"
    url: https://mifist.example.com
    email: support@mifist.example.com
servers:
  - url: https://api.asut.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
...
  /users/me/profile:
    get:
      tags:
        - Пользователи
      summary: Получить информацию о пользователе (о себе)
      description: |
        Получение профиля текущего аутентифицированного пользователя.
        
        **Требует аутентификации:** Да
        **Права доступа:** Любой аутентифицированный пользователь
        **Возвращает:** Личные данные пользователя (без проектов и ролей)
      operationId: get_user_profile
      responses:
        '200':
          description: Успешное получение профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags:
        - Пользователи
      summary: Обновить свои данные (логин, имя, фамилию)
      description: |
        Обновление личных данных пользователем.
        
        **Что можно обновить:**
        - login: Логин пользователя
        - name: Имя пользователя
        - surname: Фамилия пользователя
        
        **Ограничения:**
        - Логин должен быть уникальным в системе
        - Можно обновлять одно или несколько полей одновременно
      operationId: patch_user_profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePatchRequest'
            example:
              login: "new_login"
              name: "Иван"
              surname: "Петров"
      responses:
        '200':
          description: Данные успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePatchResponse'
        '400':
          description: Некорректный запрос (логин занят)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/me/projects:
    get:
      tags:
        - Пользователи
      summary: Получить информацию о проектах пользователя (о себе)
      description: |
        Получение проектов текущего аутентифицированного пользователя.
        
        **Требует аутентификации:** Да
        **Права доступа:** Любой аутентифицированный пользователь
        **Возвращает:** Проекты пользователя и его роли на них
      operationId: get_user_projects
      responses:
        '200':
          description: Успешное получение проектов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProjectsResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ================ Проекты ================
...
  /admin/users/{user_id}:
    get:
      tags: [Администрирование]
      summary: Получить детальную информацию о пользователе
      description: |
        Получение детальной информации о конкретном пользователе.
        
        **Требует аутентификации:** Да
        **Права доступа:** Только администраторы
        **Возвращает:** Полную информацию о пользователе и его проектах
      operationId: get_user_by_id
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID пользователя (UUID)
          example: "07f4ba61-bcdd-4fae-8db8-cf10fcbb9e72"
      responses:
        '200':
          description: Успешное получение информации о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponce'
              example:
                id: "07f4ba61-bcdd-4fae-8db8-cf10fcbb9e72"
                login: "ivan.petrov"
                surname: "Петров"
                name: "Иван"
                email: "ivan.petrov@example.com"
                status: "ACTIVE"
                is_admin: false
                created_at: "2024-01-15T10:30:00Z"
                updated_at: "2024-01-15T10:30:00Z"
                project_roles:
                  - project_id: "f9ec8db5-f1a6-4f16-aef2-a2b0e2da2f40"
                    project_name: "Банковское мобильное приложение"
                    role_id: "5bf2224d-9965-4928-8769-78fc19b014dc"
                    role_name: "PROJECT_MANAGER"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
    UserProfile:
      type: object
      required:
        - id
        - login
        - email
        - status
        - is_admin
      properties:
        id:
          type: string
          format: uuid
          description: ID пользователя
          example: "07f4ba61-bcdd-4fae-8db8-cf10fcbb9e72"
        login:
          type: string
          description: Логин пользователя
          example: "ivan.petrov"
        surname:
          type: string
          nullable: true
          description: Фамилия пользователя
          example: "Петров"
        name:
          type: string
          nullable: true
          description: Имя пользователя
          example: "Иван"
        email:
          type: string
          format: email
          description: Электронная почта пользователя
          example: "ivan.petrov@example.com"
        status:
          type: string
          description: Статус пользователя
          enum: [ACTIVE, INACTIVE, BLOCKED]
          example: "ACTIVE"
        is_admin:
          type: boolean
          description: Пользователь администратор
          example: false
    UserProfileResponse:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          description: Ответ с профилем пользователя
    UserProfilePatchRequest:
      type: object
      description: Запрос на обновление профиля пользователя
      properties:
        login:
          type: string
          nullable: true
          description: Логин пользователя
          minLength: 3
          maxLength: 50
          example: "new_login"
        name:
          type: string
          nullable: true
          description: Имя пользователя
          maxLength: 100
          example: "Иван"
        surname:
          type: string
          nullable: true
          description: Фамилия пользователя
          maxLength: 100
          example: "Петров"
    UserProfilePatchResponse:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          required:
            - message
          properties:
            message:
              type: string
              description: Сообщение об результате операции
              example: "Данные успешно обновлены"
    UserProjectRole:
      type: object
      required:
        - project_id
        - project_name
        - role_id
        - role_name
      properties:
        project_id:
          type: string
          format: uuid
          description: ID проекта
          example: "f9ec8db5-f1a6-4f16-aef2-a2b0e2da2f40"
        project_name:
          type: string
          description: Название проекта
          example: "Банковское мобильное приложение"
        role_id:
          type: string
          format: uuid
          description: ID роли
          example: "5bf2224d-9965-4928-8769-78fc19b014dc"
        role_name:
          type: string
          description: Название роли
          example: "PROJECT_MANAGER"
    UserProjectsResponse:
      type: object
      required:
        - project_roles
      properties:
        project_roles:
          type: array
          description: Роли пользователя на проектах
          items:
            $ref: '#/components/schemas/UserProjectRole'
...